pyproject.toml
[project]
name = "nyt-crossword-scraper"
version = "0.1.0"
description = "Scrape, analyze, and upload NY Times crossword comments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "torch>=2.5.0",
    "transformers>=4.46.1",
]

[tool.ruff]
# Set the maximum line length to 120.
line-length = 120

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-select = ["E501", "I"]
